check both port 80/443 configs

+++ config/scripts/post-config.d/lighttpd_custom.sh
@@ -25,11 +25,15 @@ if [ -f "/config/letsencrypt/domain.key" ] && [ ! $(stat -c %s /config/letsencry
         --exec /usr/sbin/lighttpd -- -f /etc/lighttpd/lighttpd.conf
   
   # Restart custom lighttpd webserver for custom scripts
-  sudo /sbin/start-stop-daemon --stop --pidfile /var/run/lighttpd_custom.pid
-  if [ -f "/var/run/lighttpd_custom.pid" ]; then
-    rm /var/run/lighttpd_custom.pid
+  # only if default webserver is not configured to run on port 80/443...
+  # CPO: improve: custom-server on port http-80, stock-server von https-443 should work as well!
+  if [ $(grep "https-port 443" /config/config.boot | wc -l) -eq 0 ] && [ $(grep "http-port 80" /config/config.boot | wc -l) -eq 0 ]
+    sudo /sbin/start-stop-daemon --stop --pidfile /var/run/lighttpd_custom.pid
+    if [ -f "/var/run/lighttpd_custom.pid" ]; then
+      rm /var/run/lighttpd_custom.pid
+    fi
+    sudo /sbin/start-stop-daemon --start --quiet \
+          --pidfile /var/run/lighttpd_custom.pid \
+          --exec /usr/sbin/lighttpd -- -f /config/custom/lighttpd/lighttpd_custom.conf
   fi
-  sudo /sbin/start-stop-daemon --start --quiet \
-        --pidfile /var/run/lighttpd_custom.pid \
-        --exec /usr/sbin/lighttpd -- -f /config/custom/lighttpd/lighttpd_custom.conf
 fi
